Add files via upload
The creation of the user account is now successful. Copying groups from the template and adding the new user to them also completes as expected now. Moving the new user account from the default OU to the selected OU also works now. For all intents and purposes the program works. Fro mthis point forward it will be improvements. The successful creations apply to both local and domain users. 

There needs to be a correction to allow for the middle initial. There needs to be more validation to ensure users create the account as intended (throw a popup in place to force the user to read, accept and create the account). After this has been done a feature will be added to modify existing user accounts.