Add files via upload
I colorized the label to be green when it creates the user and red when it fails. I also colorized the warning message to be red 
instead of yellow which I tried in this but it looked like garbage. I also changed the wording of the warning message that alerts 
the end user to the pre-existing SamAccountName.