Add files via upload
Started implementation of the EditUserWidget:
- Added a Load button. 
- Added a Close button
- Added a label and set text to "Select Field"
- Add a coresponding combo box for select field and fed it the contents of an array for basic users properties.
- Added a label and set text to "Select user"
- Added a coresponding combo box that prepopulates all users in the AD.

The user selects one of the following fields:
- Orginizational Unit
- Domain\UPN
- Employee Name
- Username
- Password
- Email ADdress
- Primary Proxy
- Secondary Proxy
- Display Name

Some simple checks exist to confirm the user has selected both a field and a user to review. They click the load button which drops to a switch based on the field selected. It currently reports back the value of each of the above corresponding user account attributes.